+++ spec/dummy/db/migrate/20260919000000_add_refresh_token_revoked_at_to_access_tokens.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddRefreshTokenRevokedAtToAccessTokens < ActiveRecord::Migration[6.1]
+  def change
+    add_column :oauth_access_tokens, :refresh_token_revoked_at, :datetime
+  end
+end